python_server/pyproject.toml

[project]
name = "mcp-bridge-server"
version = "0.1.0"
description = "MCP server exposing tools/resources/prompts"
requires-python = ">=3.10"
dependencies = [
  "mcp[cli]>=1.2.0",
  "python-dotenv>=1.0",
]

[tool.uv]
package = true
